refactor : 펫시터 상제 정보 조회 api - uri 변경 - 주석 삭제 - 메서드명 변경 (show -> get) - Security 설정 변경
