+++ src/main/java/com/ajou_nice/with_pet/petsitter/controller/PetSitterController.java
@@ -3,7 +3,8 @@ package com.ajou_nice.with_pet.petsitter.controller;
 
 import com.ajou_nice.with_pet.domain.dto.Response;
 import com.ajou_nice.with_pet.petsitter.model.dto.*;
-import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterDetailInfoResponse.PetSitterMyInfoResponse;
+import com.ajou_nice.with_pet.petsitter.model.dto.detail.PetSitterDetailInfoResponse;
+import com.ajou_nice.with_pet.petsitter.model.dto.detail.PetSitterDetailInfoResponse.PetSitterMyInfoResponse;
 import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.*;
 import com.ajou_nice.with_pet.petsitter.model.dto.register_info.PetSitterRegisterInfoRequest;
 import com.ajou_nice.with_pet.petsitter.model.dto.register_info.PetSitterRegisterInfoResponse;
@@ -32,17 +33,10 @@ public class PetSitterController {
 
     private final PetSitterService petSitterService;
 
-    // 펫시터에서 필요한 api //
-
-    // 펫시터 상세정보 조회 api //
-    @GetMapping("api/v1/petsitter/{petsitterId}")
+    @GetMapping("/{petSitterId}")
     @ApiOperation(value = "사용자의 펫시터 상세 정보 조회")
-    public Response<PetSitterDetailInfoResponse> showPetSitterInfo(
-            @PathVariable("petsitterId") Long petSitterId) {
-
-        PetSitterDetailInfoResponse petSitterDetailInfoResponse = petSitterService.showPetSitterDetailInfo(
-                petSitterId);
-
+    public Response<PetSitterDetailInfoResponse> getPetSitterInfo(@PathVariable("petSitterId") Long petSitterId) {
+        PetSitterDetailInfoResponse petSitterDetailInfoResponse = petSitterService.getPetSitterDetailInfo(petSitterId);
         return Response.success(petSitterDetailInfoResponse);
     }
